+++ install_and_publish.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+./gradlew clean install
+./gradlew bintrayUpload